server/requirements.txt
# SQLbot Backend Dependencies
# Locked versions for stability

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==43.0.1

# LLM / LangChain
langchain==0.3.7
langchain-core==0.3.23
langchain-openai==0.2.9
openai==1.54.5
httpx<0.28.0

# Vector store
pgvector==0.3.1

# Validation
pydantic[email]==2.10.2

# Utilities
python-dotenv==1.0.1
